feat: 랭킹 관련 api 리팩토링 및 커뮤니티 지도 작업
